df_remod

If the data frame is made of n numeric columns, a new data frame is created, with the 1st column gathering all the numeric values, and the 2nd column being the name of the columns of the initial data frame. If row names were present in the initial data frame, then a new ini_rowname column is added with the names of the rows.

If the data frame is made of one numeric column and one character or factor column, a new data frame is created, with the new columns corresponding to the split numeric values (according to the character column). NA are added a the end of each column to have the same number of rows. BEWARE: in such data frame, rows are not individuals. This means that in the example below, values 10 and 20 are associated on the same row but that means nothing in term of association.

Usage

df_remod(data, quanti.col.name = "quanti", quali.col.name = "quali")

Arguments

data

Data frame to convert.

quanti.col.name

Single character string. Optional name for the quanti column of the new data frame.

quali.col.name

Single character string. Optional name for the quali column of the new data frame.

Value

The modified data frame.

Examples

obs <- data.frame(col1 = (1:4)*10, col2 = 5:8, stringsAsFactors = TRUE) ; 
rownames(obs) <- paste0("row", 1:4) ; 
obs ;
#>      col1 col2
#> row1   10    5
#> row2   20    6
#> row3   30    7
#> row4   40    8
df_remod(obs, quanti.col.name = "quanti", quali.col.name = "quali")
#>   quanti quali ini_rowname
#> 1     10  col1        row1
#> 2     20  col1        row2
#> 3     30  col1        row3
#> 4     40  col1        row4
#> 5      5  col2        row1
#> 6      6  col2        row2
#> 7      7  col2        row3
#> 8      8  col2        row4