head2

As utils::head() but display the left or right head of big 2D objects.

Usage

head2(data1, n = 6, side = "l", safer_check = TRUE)

Arguments

data1

Any object but more dedicated for matrix, data frame or table.

n

As in head2() but for for matrix, data frame or table, number of dimension to print (10 means 10 rows and columns).

side

Either "l" or "r" for the left or right side of the 2D object (only for matrix, data frame or table).

safer_check

Single logical value. Perform some "safer" checks (see https://github.com/safer-r)? If TRUE, checkings are performed before main code running: 1) R classical operators (like "<-") not overwritten by another package because of the R scope and 2) required functions and related packages effectively present in local R lybraries. Set to FALSE if this fonction is used inside another "safer" function to avoid pointless multiple checkings.

Value

The head.

Details

BEWARE

Other arguments of head() not used.

Examples

obs1 = matrix(1:30, ncol = 5, dimnames = list(letters[1:6], LETTERS[1:5])) ; 
obs1 ; 
#>   A  B  C  D  E
#> a 1  7 13 19 25
#> b 2  8 14 20 26
#> c 3  9 15 21 27
#> d 4 10 16 22 28
#> e 5 11 17 23 29
#> f 6 12 18 24 30
head2(obs1, 3)
#>   A B  C
#> a 1 7 13
#> b 2 8 14
#> c 3 9 15