.arguments_check

Check if 1) the types of input are correct and 2) an input is missed.

Usage

.arguments_check(x, na.rm, finite, external.function.name)

Arguments

x

Numeric or logical vector or matrix or numeric table where find the initial values to calculate.

na.rm

Single logical value. Should missing values (NA and NaN) be removed ?

finite

Single logical value. Should infinite values (Inf and -Inf) be removed ? Warning: this argument does not remove NA and NaN. Please use the na.rm argument.

external.function.name

Name of the function using the .arguments_check() function.

Value

An error message if the type of input is not correct, or input missed, nothing otherwise.

Details

WARNING

arguments of the .arguments_check() function are not checked, so use carefully inside other functions

Examples

if (FALSE) {
# Example that shouldn't be run because this is an internal function
.arguments_check(x = 1:3, na.rm = TRUE)

.arguments_check(x = c(Inf, NA), na.rm = TRUE, finite = TRUE)
}