name_change

This function allow to check if a vector of character strings, like column names of a data frame, has elements present in another vector (vector of reserved words or column names of another data frame before merging).

Usage

name_change(
  data1,
  data2,
  added.string = "_modif",
  duplicate = TRUE,
  safer_check = TRUE
)

Arguments

data1

Vector of character strings to check and modify.

data2

Reference vector of character strings (duplicated elemetns not authorized).

added.string

Single character string added at the end of the modified string in elements of data1 if present in data2.

duplicate

Single logical value. Return the elements of data1 still with duplicated elements? Defaut to TRUE, i.e., duplicated elements remain duplicated.

safer_check

Single logical value. Perform some "safer" checks (see https://github.com/safer-r)? If TRUE, checkings are performed before main code running: 1) R classical operators (like "<-") not overwritten by another package because of the R scope and 2) required functions and related packages effectively present in local R lybraries. Set to FALSE if this fonction is used inside another "safer" function to avoid pointless multiple checkings.

Value

A list containing :

- $data: the modified data1 (in the same order as in the initial data1).

- $posi: the modified positions in data1. NULL if no modification.

- $ini: the initial elements before modification. NULL if no modification.

- $post: the modified elements in the same order as in ini. NULL if no modification.

Details

- If elements are duplicated in data1 and match an element of data2, then the value of added.string is added at the end of the matched element of data1. - If the same occurs with the argument duplicate = FALSE, then value of added.string is added at the end of the matched element of data1, together with an incremented number so that elements in data1 are not anymore duplicated. - An error message is returned if any of the data1 elements end with the string in added.string.

Examples

obs1 <- c("A", "B", "C", "A", "D") ; 
obs2 <- c("A", "C") ; 
name_change(obs1, obs2)
#> $data
#> [1] "A_modif" "B"       "C_modif" "A_modif" "D"      
#> 
#> $posi
#> [1] 1 3 4
#> 
#> $ini
#> [1] "A" "C" "A"
#> 
#> $post
#> [1] "A_modif" "C_modif" "A_modif"
#>